% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{assign_snp}
\alias{assign_snp}
\title{Assign snp for the subject}
\usage{
assign_snp(G1, G2, N = 1e+05, block_size = 1000)
}
\arguments{
\item{G1}{SNPs for parent1}

\item{G2}{SNPs for parent2}

\item{N}{Number of subjects}

\item{block_size}{Size of FBM to be processed}
}
\value{
Matrix of simulated SNPs
}
\description{
Assign snp for the subject
}
