% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distribution.R
\name{calc_distribution}
\alias{calc_distribution}
\title{Calculates the parameters for the conditional distribution belonging to each
liabilities of the multivariate gaussian. Of the form
mu_mult = \eqn{\Sigma _{12}\Sigma _{22}^{-1}\mu _2}
sigma_bar = \eqn{\Sigma_11 - \Sigma _{12}\Sigma _{22}^{-1}\Sigma_21}
It is used as a helper function for LTFH}
\usage{
calc_distribution(sigma)
}
\arguments{
\item{sigma}{The covariance matrix}
}
\value{
a list containing sigma_bar and mu_mult for each liability
}
\description{
Calculates the parameters for the conditional distribution belonging to each
liabilities of the multivariate gaussian. Of the form
mu_mult = \eqn{\Sigma _{12}\Sigma _{22}^{-1}\mu _2}
sigma_bar = \eqn{\Sigma_11 - \Sigma _{12}\Sigma _{22}^{-1}\Sigma_21}
It is used as a helper function for LTFH
}
\examples{
calc_distribution(get_cov(0.5))
}
