% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pheno_liabil.R
\name{calc_full_liabil}
\alias{calc_full_liabil}
\title{Calculate full liabilities from envirromental and genetic liabilities}
\usage{
calc_full_liabil(liabilities, n_sib = 0)
}
\arguments{
\item{liabilities}{Liabilities of subject and family}

\item{n_sib}{Amount of siblings}

\item{K}{The prevalance of trait}
}
\value{
Matrix containing full lliabilities for each subject
}
\description{
Calculate full liabilities from envirromental and genetic liabilities
}
