% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pheno_liabil.R
\name{calc_phen}
\alias{calc_phen}
\title{Calculate phentypes from from envirromental and genetic liabilities}
\usage{
calc_phen(liabilities, K = 0.05, n_sib = 0)
}
\arguments{
\item{liabilities}{Liabilities of subject and family}

\item{K}{The prevalance of trait}

\item{n_sib}{Amount of siblings}
}
\value{
Matrix containing phenotypes for each subject
}
\description{
Calculate phentypes from from envirromental and genetic liabilities
}
