% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{collapse_data}
\alias{collapse_data}
\title{Calculates the sum of colunms in a tibble for each person and value}
\usage{
collapse_data(data, n_sib)
}
\arguments{
\item{data}{Tibble with colunms to sum over,}

\item{n_sibs}{Number of siblings}
}
\value{
Collapsed tibble
}
\description{
Calculates the sum of colunms in a tibble for each person and value
}
