% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{control_plot}
\alias{control_plot}
\title{Plotting estimated liabilities agianst distribution}
\usage{
control_plot(phenos, h2, col = "black")
}
\arguments{
\item{h2}{The heritability}

\item{pheno}{matrix containing phenotypes}
}
\value{
Plot of estimated liabilities
}
\description{
Plotting estimated liabilities agianst distribution
}
