% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{create_fbm}
\alias{create_fbm}
\title{Create and save FBM}
\usage{
create_fbm(filename, N, M)
}
\arguments{
\item{filename}{Filename for FBM}

\item{N}{Amount of individuals}

\item{M}{Amount of SNPs}
}
\value{
A FBM
}
\description{
Create and save FBM
}
