% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{decision_cross}
\alias{decision_cross}
\title{Creates evaluation plot for decision boundary}
\usage{
decision_cross(
  train_data,
  y,
  cross_folds,
  bounds,
  thr,
  ncores = 1,
  LogReg = FALSE
)
}
\arguments{
\item{train_data}{List generated from gen_sim.}

\item{y}{The target vector. Could either be estimated liabilities from LTFH or phenotypes.}

\item{cross_folds}{Number of folds in cross validation.}

\item{bounds}{Decision boundaries to plot outcome for.}

\item{thr}{Threshold for p-value to be used in calculating PRS.}

\item{ncores}{Amount of cores to be used.}

\item{LogReg}{Boolean indicating if logistic regression should be used to estimate the casual effect.}
}
\value{
The confusion matrix for the mean of each bound.
}
\description{
The function creates one plot for each given bound. It shows the confusion matrix for the mean of each bound.
The function can be used to find which boundary for PRS best describes the data.
}
