% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_check.R
\name{dist_check}
\alias{dist_check}
\title{Check distribution of liabilities}
\usage{
dist_check(data)
}
\arguments{
\item{data}{List generated from gen_sim.}
}
\value{
Creates a qqplot, mean and sample variance for the liability of each family.
}
\description{
Check distribution of liabilities
}
\examples{
dist_check(snp_attach("test.rds"))
}
