% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_LTFH.R, R/estimate_liabilities.R
\name{estimate_liabilities}
\alias{estimate_liabilities}
\title{Estimates liabilities for every subject}
\usage{
estimate_liabilities(pheno, K = 0.05, n_sib = 0, h2 = 0.5)

estimate_liabilities(pheno, K = 0.05, n_sib = 0, h2 = 0.5)
}
\arguments{
\item{pheno}{matrix containing phenotypes}

\item{K}{The prevalence of trait}

\item{n_sib}{Amount of siblings}

\item{h2}{The heritability}
}
\value{
Estimated liabilities

Tibble containing phenotypes and estimates of the individual and family
}
\description{
Estimates liabilities for every subject

Using phentoypes the function estimates full liabilities using LT-FH
}
