% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{G_func_simple}
\alias{G_func_simple}
\title{Simulate SNP for all subjects}
\usage{
G_func_simple(filename, MAF, N = 1e+05, M = 1e+05, block_size = 1000)
}
\arguments{
\item{filename}{Filename for fbm}

\item{MAF}{Probabilities of mutaion}

\item{N}{number of subjects}

\item{block_size}{Size of fbm to be processed}

\item{n_sibs}{Number of siblings}
}
\value{
Simulated SNPs for all subjects
}
\description{
Simulate SNP for all subjects
}
