% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_sim}
\alias{gen_sim}
\title{Simulate genetic data}
\usage{
gen_sim(
  filename,
  N = 1e+05,
  M = 1e+05,
  n_sib = 0,
  K = 0.05,
  h_sq = 0.5,
  C = 1000,
  block_size = 1000,
  fam = TRUE
)
}
\arguments{
\item{filename}{Filename for fbm}

\item{N}{number of subjects}

\item{M}{Number of SNPs}

\item{K}{The prevalance of trait}

\item{C}{Amount of causal SNP's}

\item{block_size}{Size of fbm to be processed}

\item{fam}{Boolean deciding if simulation should include family structure}

\item{h2}{The heritability}

\item{n_sibs}{Number of siblings}
}
\value{
bigsnp object containing fbm, information about subject and family for liabilities, and generated SNP info
}
\description{
Simulate genetic data
}
