% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{get_member}
\alias{get_member}
\title{Simulate block of snps for each family member}
\usage{
get_member(i, beta, MAF, N = 1e+05, block_size = 1000)
}
\arguments{
\item{i}{Iteration from for loop.}

\item{beta}{A vector containing the casual effect of each SNP.}

\item{MAF}{A vector containing minor allele frequencies.}

\item{N}{Amount of subjects.}

\item{block_size}{Size of FBM to be processed in each iteration.}
}
\value{
List containing block of simulated snps and genetic liabilities
}
\description{
Helper function for gen_sim()
}
