% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.R
\name{GWAS}
\alias{GWAS}
\title{Computes casusal SNP's}
\usage{
GWAS(G, y, p, logreg = FALSE, ncores = 1)
}
\arguments{
\item{G}{A file backed matrix with SNP data.}

\item{y}{A vector containing targets for each position.}

\item{p}{The significance level.}

\item{logreg}{if TRUE do logistic regression, else Linear regression.}

\item{ncores}{Amount of cores to use.}
}
\value{
A matrix containing: \cr
\itemize{
\item The slopes of each regression \cr
\item The standard errors of each slope \cr
\item The t-scores associated with each slope. \cr
\item The p-values for each slope \cr
\item And a binary vector saying if the position is causal or not given p
}
}
\description{
This function is heavily build onto bigstatsr::big_univLinReg. For more documentation check \cr https://privefl.github.io/bigstatsr/reference/big_univLinReg.html
}
