% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{liabilities_func_fam}
\alias{liabilities_func_fam}
\title{Calculate genetic liabilities and simulate enviromental liabilities for subject and family}
\usage{
liabilities_func_fam(
  G,
  beta,
  MAF,
  liab,
  N = 1e+05,
  n_sib = 0,
  K = 0.05,
  h2 = 0.5,
  block_size = 1000
)
}
\arguments{
\item{G}{A FBM containing SNP data}

\item{beta}{A vector containing the casual SNPs}

\item{MAF}{A vector containing minor allele frequencies}

\item{liab}{A tibble containing genetic liabilities for each family member}

\item{N}{Amount of individuals}

\item{n_sib}{Amount of siblings}

\item{block_size}{the size of each iteration}

\item{h_sq}{heritability}
}
\value{
A tibble containing genetic and enviromental liability for each family member
}
\description{
Calculate genetic liabilities and simulate enviromental liabilities for subject and family
}
