% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{liabilities_func_simple}
\alias{liabilities_func_simple}
\title{Calculate genetic liabilities and simulate enviromental liabilities for subject}
\usage{
liabilities_func_simple(
  G,
  beta,
  MAF,
  N = 1e+05,
  K = 0.05,
  h2 = 0.5,
  block_size = 1000
)
}
\arguments{
\item{G}{File backed matrix containing SNP data.}

\item{beta}{A vector containing the casual effect of each SNP.}

\item{MAF}{A vector containing minor allele frequencies.}

\item{N}{Amount of subjects.}

\item{h2}{The heritability of trait.}

\item{block_size}{Size of FBM to be processed in each iteration.}
}
\value{
A tibble containing genetic and full liability for each subject.
}
\description{
Helper function for gen_sim()
}
