% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFH_notrunc.R
\name{LTFH_notrunc}
\alias{LTFH_notrunc}
\title{Create LTFH estimations without truncation.}
\usage{
LTFH_notrunc(covmat, N = 3000, K = 0.05, s_val = 0)
}
\arguments{
\item{covmat}{The covariance matrix}

\item{N}{Amount of rows in the matrix that the estimations are saved in}

\item{K}{The prevalance of trait}

\item{s_val}{The starting value of estimates}
}
\value{
A vector containing LTFH estimate of liabilities of the form \cr
c(genetic_liability_subject, liability_subject, liability_parent1, \cr
liability_parent2, liability_sibling1, ..., liability_siblingN)
}
\description{
Create LTFH estimations without truncation.
}
\examples{
LTFH(get_cov(0.5, n_sib = 1))
}
