% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{MAF_func}
\alias{MAF_func}
\title{Simulate vector containing minor allele frequencies (MAF)}
\usage{
MAF_func(M = 1e+05)
}
\arguments{
\item{M}{Amount of SNPs.}
}
\value{
Vector containing minor allele frequencies (MAF).
}
\description{
The function simulates minor allele frequencies for each SNP.
Minor allele frequencies is independent and uniformly distributed on the interval \eqn{[0.01, 0.49]}.
}
