% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Creates a Manhattan plot}
\usage{
manhattan_plot(gwas_summary, beta, thresholds = FALSE)
}
\arguments{
\item{gwas_summary}{Output from GWAS.}

\item{beta}{A vector containing the actual casual effect of each SNP.}

\item{thresholds}{A vector containing p-values. The function draws a horisontal
line at the given values.}
}
\value{
A Manhattan plot.
}
\description{
This function returns a Manhattan plot.
}
