% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{no_trunc_estimate}
\alias{no_trunc_estimate}
\title{Returns random value from normal distribution.}
\usage{
no_trunc_estimate(mu, sigma)
}
\arguments{
\item{mu}{The mean.}

\item{sigma}{The variance.}
}
\value{
Random value from normal distribution.
}
\description{
Returns random value from normal distribution.
}
