% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{normalized_prod}
\alias{normalized_prod}
\title{Normalize file backed matrix (FBM)}
\usage{
normalized_prod(Gx, beta, MAF)
}
\arguments{
\item{Gx}{Subset of FBM containing SNP data.}

\item{beta}{A vector containing the casual effect of each SNP.}

\item{MAF}{A vector containing minor allele frequencies.}
}
\value{
A vector containing the genetic liability for Gx.
}
\description{
Normalize file backed matrix (FBM)
}
