% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{plot_gibbs}
\alias{plot_gibbs}
\title{Plot the iterations of gibb_sampl}
\usage{
plot_gibbs(ests)
}
\arguments{
\item{ests}{output from gibb_sampl}
}
\value{
plots showing values at each iteration
}
\description{
Plot the iterations of gibb_sampl
}
