% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{power_plot}
\alias{power_plot}
\title{Create power plot}
\usage{
power_plot(gwas_summary, beta)
}
\arguments{
\item{gwas_summary}{output from GWAS}

\item{beta}{A vector containing the actual casual effect of each SNP.}
}
\value{
A power plot
}
\description{
Create power plot
}
