% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_model}
\alias{pred_model}
\title{Trains model on train data given the target vector.}
\usage{
pred_model(train_data, y, thr, ncores = 1, LogReg = FALSE)
}
\arguments{
\item{train_data}{List generated from gen_sim.}

\item{y}{The target vector. Could either be estimated liabilities from LTFH or phenotypes.}

\item{thr}{Threshold for p-value to be used in calculating PRS.}

\item{ncores}{Amount of cores to be used.}

\item{LogReg}{Boolean indicating if logistic regression should be used to estimate the casual effect.}
}
\value{
List containing output from GWAS and Linear regression of PRS on phenotype of the subject.
}
\description{
The function will use GWAS to find the casual effect of each SNP.
It will then calculate the PRS.
The final output is list containing the model from GWAS. This can be used to predict the phenotypes of test data.
}
