% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_cross.R
\name{PRS_cross}
\alias{PRS_cross}
\title{Uses cross validation to train and estimate PRS for each fold}
\usage{
PRS_cross(data, y01, cross_folds, LogReg = FALSE)
}
\arguments{
\item{data}{List generated from gen_sim.}

\item{cross_folds}{Number of folds in cross validation.}

\item{LogReg}{Boolean indicating if logistic regression should be used to estimate the casual effect.}

\item{y}{The target vector. Could either be estimated liabilities from LTFH or phenotypes.}
}
\value{
List with estimated PRS for each fold.
}
\description{
Uses cross validation to train and estimate PRS for each fold
}
