% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{prs_plot}
\alias{prs_plot}
\title{Creates evaluation plot for MSE, AUC or r squared}
\usage{
prs_plot(PRS, data, method = "MSE")
}
\arguments{
\item{PRS}{Output from PRS_cross.}

\item{data}{List generated from gen_sim.}

\item{method}{Indicates Which method each fold should be evaluated on.}
}
\value{
List containing two plots, one for each threshold and fold and one mean of all folds.
}
\description{
The function creates one plot for each fold in PRS.
This plot will show the chosen method evaluated for each threshold.
Lastly the function will plot the mean for each threshold.
The function can be used to find which threshold. of p-values best can describe the data.
}
