% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Plots true causal effects agianst estimated effects}
\usage{
scatter_plot(gwas_summary, beta)
}
\arguments{
\item{gwas_summary}{Output from GWAS.}

\item{beta}{A vector containing the actual casual effect of each SNP.}
}
\value{
Scatterplot of true causal effects agianst estimated effects.
}
\description{
Plots true causal effects agianst estimated effects
}
