% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_attach.R
\name{snp_attach}
\alias{snp_attach}
\title{Attach rds_file}
\usage{
snp_attach(rds_file)
}
\arguments{
\item{rds_file}{An rds file}
}
\value{
The values from the rds file.
}
\description{
For ease of use this function is added
It is the exact same as bigsnpr::snp_attach see ?bigsnpr::snp_attach for further documentation
}
