% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{trunc_estimate}
\alias{trunc_estimate}
\title{Returns truncated random value from normal distribution.}
\usage{
trunc_estimate(j, phenos, K = 0.05, mu, sigma)
}
\arguments{
\item{j}{Index for familymember}

\item{phenos}{Phenotypes for which estimates of liabilities is wanted}

\item{K}{The prevalance of trait}

\item{mu}{The mean}

\item{sigma}{The variance}
}
\value{
Value
}
\description{
Returns truncated random value from normal distribution.
}
