% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_func.R
\name{y_func}
\alias{y_func}
\title{Calculate phenotypes}
\usage{
y_func(l_e_0, l_g_0, K = 0.05)
}
\arguments{
\item{l_e_0}{vector containing enviromental liabilities}

\item{l_g_0}{vector containing genetic liabilities}

\item{k}{prevalance of trait}
}
\value{
binary vector contatining phenotypes
}
\description{
Calculate phenotypes
}
