% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{beta_func}
\alias{beta_func}
\title{Simulate beta}
\usage{
beta_func(M = 1e+05, h2 = 0.5, C = 1000)
}
\arguments{
\item{M}{size of the beta vector.}

\item{h2}{The heritability of trait.}

\item{C}{Amount of causal SNPs.}
}
\value{
A vector of size M containing a value from rnorm(1, 0, sqrt(h2/C)) at C random places.
}
\description{
Simulate beta
}
