% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{beta_func}
\alias{beta_func}
\title{Simulate beta}
\usage{
beta_func(C = 1000, h_sq = 0.5, M = 1e+05)
}
\arguments{
\item{C}{Amount of causal SNPs}

\item{M}{size of the beta vector}

\item{h2}{The heritability}
}
\value{
A vector of size m containing a value from rnorm(1, 0, sqrt(h_sq/C)) \cr
a C random places.
}
\description{
Simulate beta
}
