% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{control_plot}
\alias{control_plot}
\title{Plotting estimated liabilities agianst distribution}
\usage{
control_plot(phenos, h2, col = "black")
}
\arguments{
\item{h2}{The heritability of trait.}

\item{pheno}{Vector containing phenotypes.}
}
\value{
Plot of estimated liabilities.
}
\description{
A recreation of figure 1. b in Hujoel et al. (2020)
}
