% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{create_fbm}
\alias{create_fbm}
\title{Create and save file backed matrix (FBM)}
\usage{
create_fbm(filename, N, M)
}
\arguments{
\item{filename}{Filename for FBM and rds file.}

\item{N}{Amount of subjects.}

\item{M}{Amount of SNPs.}
}
\value{
A FBM.
}
\description{
Helper function for gen_sim()
}
