% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_check.R
\name{dist_check}
\alias{dist_check}
\title{Check distribution of liabilities}
\usage{
dist_check(data)
}
\arguments{
\item{data}{List generated from gen_sim.}
}
\value{
Creates a qqplot, mean and sample variance for the liability of each family.
}
\description{
The function calculates the mean and variance of each full liability returned by gen_sim().
The assumption which the simulation is based on is that each full liability will be normaly distributed.
The function therefore also returns the qqplot of each full liability. The purpose of this function is to provide
certainty that the simulated data are correct.
}
\examples{
dist_check(snp_attach("genetic_data.rds"))
}
