% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{G_func_simple}
\alias{G_func_simple}
\title{Simulate SNP for all subjects}
\usage{
G_func_simple(filename, MAF, N = 1e+05, M = 1e+05, block_size = 1000)
}
\arguments{
\item{filename}{Filename for file backed matrix(FBM) and rds file.}

\item{MAF}{A vector containing minor allele frequencies.}

\item{N}{Amount of subjects.}

\item{block_size}{Size of FBM to be processed in each iteration.}

\item{n_sib}{Amount of siblings.}
}
\value{
FBM containing simulated SNPs for all subjects.
}
\description{
The function creates a bigsnpr object, saves it in a rds file.
and fills the file backed matrix generated with geneotypes for each subject.
}
