% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_LTFH.R, R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Creates tibble containing information and estimates for every subject}
\usage{
get_all_data(data, K = 0.05, n_sib = 0, h2 = 0.5)

get_all_data(data, K = 0.05, n_sib = 0, h2 = 0.5)
}
\arguments{
\item{data}{Tibble containing information about each indiviual and family's genetic and enviromental liabilities}

\item{K}{The prevalence of trait}

\item{n_sib}{Amount of siblings}

\item{h2}{The heritability}
}
\value{
Tibble with data and estimates

Tibble containing all information and estimates of the individual and family
}
\description{
Creates tibble containing information and estimates for every subject

Using genetic and enviromental liabilities the function calculates phenotypes, full liabilites and estimates full liabilities using LT-FH
}
