% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov.R
\name{get_cov}
\alias{get_cov}
\title{Create covariance matrix}
\usage{
get_cov(h2, n_sib = 0)
}
\arguments{
\item{h2}{The heritability of trait.}

\item{n_sib}{Amount of siblings.}
}
\value{
Covariance matrix of datatype matrix.
}
\description{
The function returns the covariance matrix for the multivariate normal distribution of ()
}
\examples{
get_cov(0.5, n_sib = 2)
}
