% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{get_index}
\alias{get_index}
\title{Find start and and end index of FBM}
\usage{
get_index(i, block_size)
}
\arguments{
\item{i}{Iteration from for loop.}

\item{block_size}{Size of FBM to be processed in each iteration.}
}
\value{
List of start and and end index.
}
\description{
Helper function for gen_sim()
}
