% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Adds a postfix to given names for each family member}
\usage{
get_names(names, n_sib = 0, id = TRUE, parents = TRUE)
}
\arguments{
\item{names}{Names of the the values to add a postfix to e.g. l_g.}

\item{n_sib}{Amount of siblings.}

\item{id}{If TRUE add a postfix for the subject.}

\item{parents}{If TRUE add postfix for parents.}
}
\value{
A vector containing the names for each family member.
}
\description{
This function serves as a help for the user to generate the wished columns names.
The function works well in combination with dplyr using the argument contains in select.
There is no requirements for input other than it's a string.
}
\examples{
get_names(c("l_g", "l_e"), n_sib = 2)
}
