% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.R
\name{GWAS}
\alias{GWAS}
\title{Computes casusal SNP's}
\usage{
GWAS(G, y, ncores = 1, p = 0.05/1e+06)
}
\arguments{
\item{G}{A FBM with SNP data}

\item{y}{A vector containing targets for each position.}

\item{ncores}{Amount of cores to use}

\item{p}{The significance level standard is 0.05 with bonferroni correction}
}
\value{
a matrix containing the slopes of each regression, \cr
the standard errors of each slope \cr
the t-scores associated with each slope. \cr
the p-values for each slope \cr
and a binary vector saying if the position is causal or not given p
}
\description{
This function is heavily build onto bigstatsr::big_univLinReg. For more documentation check \cr https://privefl.github.io/bigstatsr/reference/big_univLinReg.html
}
