% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{liabilities_func_fam}
\alias{liabilities_func_fam}
\title{Calculate genetic liabilities and simulate enviromental liabilities for subject and family}
\usage{
liabilities_func_fam(
  G,
  beta,
  MAF,
  liab,
  N = 1e+05,
  n_sib = 0,
  K = 0.05,
  h2 = 0.5,
  block_size = 1000
)
}
\arguments{
\item{G}{File backed matrix containing SNP data.}

\item{beta}{A vector containing the casual effect of each SNP.}

\item{MAF}{A vector containing minor allele frequencies.}

\item{liab}{A tibble containing genetic liabilities for each subject's family members.}

\item{N}{Amount of subjects.}

\item{n_sib}{Amount of siblings.}

\item{block_size}{Size of FBM to be processed in each iteration.}

\item{h_2}{The heritability of trait.}
}
\value{
A tibble containing genetic and full liability for each family member.
}
\description{
Helper function for gen_sim()
}
