% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{liabilities_func_simple}
\alias{liabilities_func_simple}
\title{Calculate genetic liabilities and simulate enviromental liabilities for subject}
\usage{
liabilities_func_simple(
  G,
  beta,
  MAF,
  N = 1e+05,
  K = 0.05,
  h2 = 0.5,
  block_size = 1000
)
}
\arguments{
\item{G}{A FBM containing SNP data}

\item{beta}{a vector containing the casual SNPs}

\item{MAF}{A vector containing minor allele frequencies}

\item{N}{Amount of individuals}

\item{block_size}{the size of each iteration}

\item{h_sq}{heritability}
}
\value{
A tibble containing genetic and enviromental liability for each subject
}
\description{
Calculate genetic liabilities and simulate enviromental liabilities for subject
}
