% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{liabilities_func_simple}
\alias{liabilities_func_simple}
\title{Calculate genetic liabilities and simulate enviromental liabilities for subject}
\usage{
liabilities_func_simple(G, MAF, beta, N = 1e+05, h_sq = 0.5, block_size = 1000)
}
\arguments{
\item{G}{A FBM containing SNP data}

\item{MAF}{a vector containing minor allele frequencies}

\item{beta}{Causal snps}

\item{N}{Amount of individuals}

\item{h_sq}{heritability}

\item{block_size}{the size of each iteration}
}
\value{
A tibble containing genetic and enviromental liability for each subject
}
\description{
Calculate genetic liabilities and simulate enviromental liabilities for subject
}
