% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_LTFH.R
\name{LTFH}
\alias{LTFH}
\title{Estimates liabilities for every subject}
\usage{
LTFH(data, n_sib = 0, K = 0.05, h2 = 0.5)
}
\arguments{
\item{data}{Family data}

\item{n_sib}{Amount of siblings}

\item{K}{The prevalance of trait}

\item{h2}{The heritability}
}
\value{
Tibble containing input data and estimated liabilities
}
\description{
Estimates liabilities for every subject
}
