% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_LTFH.R
\name{LTFH}
\alias{LTFH}
\title{Estimates liabilities for every subject}
\usage{
LTFH(data, n_sib = 0, K = 0.05, h2 = 0.5)
}
\arguments{
\item{data}{List generated from gen_sim.}

\item{n_sib}{Amount of siblings.}

\item{K}{The prevalance of trait.}

\item{h2}{The heritability of trait.}
}
\value{
Tibble containing data$fam and estimated liabilities.
}
\description{
Estimates liabilities for every subject
}
