% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFH.R
\name{LTFH}
\alias{LTFH}
\title{Create LTFH estimations}
\usage{
LTFH(
  covmat,
  phenos,
  N = 10000,
  K = 0.05,
  s_val = 0,
  min_run = 400,
  all_est = FALSE
)
}
\arguments{
\item{covmat}{The covariance matrix}

\item{phenos}{A binary vector containing the phenotype for each family member
of the form \cr c(p_subject, p_parent1, p_parent2, p_sibling1, ... ,p_siblingN) \cr
where p_familymember is a binary value (1 or 2) \cr}

\item{N}{Amount of rows in the matrix that the estimations are saved in}

\item{K}{The prevalance of trait}

\item{s_val}{The starting value of estimates}

\item{min_run}{The minimum amount of runs to do after burn in}

\item{all_est}{If TRUE return the value for each iteration after burn in}
}
\value{
A vector containing LTFH estimate of liabilities of the form \cr
c(genetic_liability_subject, liability_subject, liability_parent1, \cr
liability_parent2, liability_sibling1, ..., liability_siblingN)
}
\description{
Create LTFH estimations
}
\examples{
LTFH(get_cov(0.5, n_sib = 1), c(1, 1, 0, 0))
}
