% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_LTFH.R
\name{LTFH}
\alias{LTFH}
\title{Estimates liabilities for every subject}
\usage{
LTFH(data, n_sib = 0, K = 0.05, h2 = 0.5)
}
\arguments{
\item{data}{List generated from gen_sim.}

\item{n_sib}{Amount of siblings.}

\item{K}{The prevalance of trait.}

\item{h2}{The heritability of trait.}
}
\value{
Tibble containing data$fam and estimated liabilities.
}
\description{
This function takes the output from gen_sim and uses the phenotypes to estimate the genetic liability for each subject.
The function estimated the liability for each configuration of phenotypes and joins these values on the existing dataframe.
Subjects having the same configuration will therefore have the same estimated genetic liability.
}
