% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{MAF_func}
\alias{MAF_func}
\title{Simulate vector containing minor allele frequencies}
\usage{
MAF_func(M = 1e+05)
}
\arguments{
\item{M}{Number of SNPs}
}
\value{
Vector containing minor allele frequencies
}
\description{
Simulate vector containing minor allele frequencies
}
