% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Creates a manhattan plot}
\usage{
manhattan_plot(gwas_summary, beta, thresholds = FALSE)
}
\arguments{
\item{gwas_summary}{Output from GWAS.}

\item{beta}{A vector containing the actual casual effect of each SNP.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot
}
