% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{normalized_prod}
\alias{normalized_prod}
\title{Normalize FBM}
\usage{
normalized_prod(Gx, beta, MAF)
}
\arguments{
\item{Gx}{small matrix containing snp data}

\item{probs}{Probabilities of mutaion}

\item{beta_cut}{Causal SNPs}
}
\value{
Matrix containing normalized values
}
\description{
Normalize FBM
}
