% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{plot_gibbs}
\alias{plot_gibbs}
\title{Plot the convergence of gibb_sampl}
\usage{
plot_gibbs(ests)
}
\arguments{
\item{ests}{Output from gibb_sampl with all_est=TRUE.}
}
\value{
Plots that shows values at each iteration.
}
\description{
Plot the convergence of gibb_sampl
}
