% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulisations.R
\name{power_plot}
\alias{power_plot}
\title{Creates a power plot}
\usage{
power_plot(gwas_summary, beta)
}
\arguments{
\item{gwas_summary}{Output from GWAS.}

\item{beta}{A vector containing the actual casual effect of each SNP.}
}
\value{
A power plot.
}
\description{
Creates a plot with values of estimates on the x-axis and power of estimates below the values on the x-axis.
}
