% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predicts the probability of having the given trait}
\usage{
prediction(test_data, gwas, thr)
}
\arguments{
\item{test_data}{List generated from gen_sim.}

\item{thr}{Threshold for p-value to be used in calculating PRS.}

\item{model}{Output from pred_model.}
}
\value{
A matrix containing predictions.
}
\description{
The function will take outcome from pred_model and use it an to find the PRS of each subject having the trait/sickness.
}
