% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predicts the probability of having the given trait}
\usage{
prediction(test_data, model, thr)
}
\arguments{
\item{test_data}{List generated from gen_sim.}

\item{model}{Output from pred_model().}

\item{thr}{Treshold for p-value to be used in calculating PRS.}
}
\value{
Vector of estimated probabilities.
}
\description{
The function will take outcome from pred_model() and use it an to find the estimated probability of each subject having the trait/sickness.
}
